move spotify to auth endpoints
