+++ app/src/app/Nav.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren } from 'react';
 import { AppBar, Box, Toolbar, Typography, Link, Stack } from '@mui/material';
 import { useCurrentUser } from '../hooks/use-current-user';
-import { Playbar } from './playbar';
+import { AuthLinks } from '../lib/auth';
 
 function NavProfileDisplay(
   props: PropsWithChildren<{
@@ -56,10 +56,10 @@ export function Nav() {
             <NavLink href="https://accounts.spotify.com/en/status">
               Account
             </NavLink>
-            <NavLink href="/spotify/logout">Logout</NavLink>
+            <NavLink href={AuthLinks.logout}>Logout</NavLink>
           </Stack>
         ) : (
-          <NavLink href="/spotify/login">Login</NavLink>
+          <NavLink href={AuthLinks.login}>Login</NavLink>
         )}
       </Toolbar>
     </AppBar>
