+++ app/src/lib/auth.ts
@@ -1,0 +1,4 @@
+export const AuthLinks = {
+  login: '/auth/login',
+  logout: '/auth/logout',
+};
