+++ app/src/setupProxy.js
@@ -1,13 +1,6 @@
 const { createProxyMiddleware } = require('http-proxy-middleware');
 
 module.exports = function (app) {
-  // app.use(
-  //   '/api',
-  //   createProxyMiddleware({
-  //     target: 'http://localhost:8080',
-  //     changeOrigin: true,
-  //   })
-  // );
   app.use(
     '/api',
     createProxyMiddleware({
@@ -17,6 +10,13 @@ module.exports = function (app) {
   );
   app.use(
     '/spotify',
+    createProxyMiddleware({
+      target: 'http://localhost:8080',
+      changeOrigin: true,
+    })
+  );
+  app.use(
+    '/auth',
     createProxyMiddleware({
       target: 'http://localhost:8080',
       changeOrigin: true,
