+++ server/.env
@@ -3,4 +3,4 @@ CLIENT_BASE_URL=http://localhost:3000
 
 SPOTIFY_API_CLIENT_ID=19fe4b9ac3ad450fa1ef5a7f5934f0be
 SPOTIFY_API_CLIENT_SECRET=3752b7c89d0748a5bd49855fbdf4d658
-SPOTIFY_API_REDIRECT_URI=http://localhost:8080/spotify/callback
+SPOTIFY_API_REDIRECT_URI=http://localhost:8080/auth/callback
